+++ aquadoggo/src/db/mod.rs
@@ -3,7 +3,7 @@
 //! Persistent storage for an `aquadoggo` node supporting both Postgres and SQLite databases.
 //!
 //! The main interface is [`SqlStore`] which offers an interface onto the database by implementing
-//! the storage traits defined in `p2panda-rs` as well as some implementation specific features.  
+//! the storage traits defined in `p2panda-rs` as well as some implementation specific features.
 use anyhow::{Error, Result};
 use sqlx::any::{Any, AnyPool, AnyPoolOptions};
 use sqlx::migrate;
@@ -11,6 +11,7 @@ use sqlx::migrate::MigrateDatabase;
 
 pub mod errors;
 pub mod models;
+pub mod query;
 pub mod stores;
 pub mod types;
